id to Long

+++ src/main/java/com/yana/internship/repository/ApartmentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApartmentRepository extends CrudRepository<Apartment, Integer> {
+public interface ApartmentRepository extends CrudRepository<Apartment, Long> {
 }
