+++ src/main/java/com/yana/internship/repository/FileRepository.java
@@ -3,5 +3,5 @@ package com.yana.internship.repository;
 import com.yana.internship.entity.File;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FileRepository extends CrudRepository<File, Integer> {
+public interface FileRepository extends CrudRepository<File, Long> {
 }
